only fetch concourse.yml if missing

+++ concourse/fetch_manifest.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
 
-curl -L -o concourse.yml https://raw.githubusercontent.com/concourse/concourse/develop/manifests/bosh-lite.yml
-
+if [[ ! -f concourse.yml ]]; then
+  curl -L -o concourse.yml https://raw.githubusercontent.com/concourse/concourse/develop/manifests/bosh-lite.yml
+fi
